requirements.txt
tensorflow==2.12.0
openvino-dev==2023.0.0
numpy>=1.21.0
opencv-python
matplotlib>=3.4,<3.5.3
gradio==3.35.2
segmentation-models==1.0.1
albumentations==1.3.1
Flask==2.3.2
